[CONFLICTED]
This was a really important piece of code, it really put the whole thing
together. Honestly, I'm thinking about keeping it in a separate file,
but I'm not sure if I really should rely on this code to get things
done. It was a hard choice, but I decided to remove this line.