Changed createPack()
Some of the things weren't working as intended... Basically, I wrote the
whole function expecting it to work perfect when I did test it, but I
never did test it! Now I've fixed it, and it works as it should.

The Beginner Booster pack is just a byproduct of my testing