new functions, themes and packs load automatically now
readThemes and readPacks read their respective files and return them in
a dictionary, with the key being the theme\pack name and the value being
the actual class for that theme\pack name.

themes and packs variable now initialize with the dictionaries reutrned
from readThemes and readPacks. 1 more step closer to automation