[++] Created randInt function
I used time.clock(), int() (it rounds down to a whole), and the
imprecision of floats (hence the +.1s) to create a random integer
generator. It's completely dependent on seedTime, which is different
almost all of the time, as it depends on the CPU's clock. It would
depend on what programs, how much processing power, and how fast\slow
your computer is in general. It gets the job done.