[++] Added some clarity things + more tests
Now have 3 different tests. Notice how the speed greatly differs, but
results are (basically) the same. 1 test with 10000 single draws, 1 test
with 1000 10-item draws, and 1 test with 100 100-item draws. The 10000
one goes much slower, but it's perfectly fine for doing 1-draw tasks. If
you want to go with bigger amounts based on the same inputs and weights,
it's better to do all of the draws in the same function call, rather
than x different calls to the function

Added a perfect results dict to compare the RNG against perfection.