[+-] Changed makeCards and openPack in Pack class
Changed makeCard to makeCards. Now you don't have to call on that
function multiple times in a loop, just pass a different amount (above
0) through the cardAmt and it will generate that many cards for you.

Changed openPack to fit with the new makeCards. It now (possibly)
generates random themed cards AFTER generating the pack, and replaces a
random card in the pack with the new themed card. This way makes it much
more rare, as it only generates the chance (maxThemed) times, instead of
(cardAmt) times, so if you have a 5% chance for 1 themed card, you'll
have a (5%*5%) chance for 2 themed cards (assuming your max themed card
limit is 2)