[+-] Changed weightchoice functionality
Now instead of inputting a list and it generating an even weightList
based on how many inputs there are and using tons of memory with popList
(e.g. 500 items on list, each weight on weightList would be 1/500, and
there would be 500 of each item on popList), it now just puts the
inputList into popList. It is literally the same exact thing, but it
uses much less memory and does NOT cause annoying floating point errors.

Also added a result variable in weightChoice so that it is more readable
by returning\appending result rather than returning\appending
popList[~...]