[+-] Added testPack(), changed openPack()
added testPack() to see if openPack() was properly generating close to
the themeCardChance percentage amount of cards, but it turns out it
averaged at around 1/3 of the expected amount, so I added these
modifiers to raise that average percentage amount closer to the optimal
amount.

The goal here was to hit about 70-80% of the expected amount to give
more rarity to the cards, while still having more than double of
previous iterations.