[!!] Good info right here
Dev with the TCG itself will stop for maybe a few days, because I am
working on an account system that lets you create and verify your
account. It will be used to login and access your TCG account. You will
be able to buy card packs, view your inventory of cards, view your
currency amount and much more. I will be doing all of this before
creating the GUI, because that will make the GUI development much more
simple and straightforward, rather than adding something and modifying
(possibly) all the code so that it works. It will need to work with it
all on release, rather than having to "patch" it up as I create new
things.