[++] simple email sending module i made
Pretty much exactly what it seems. Instead of all that confusing stuff
every time I want to send an email, just input the things and it does it
for you! That's what making functions is for!

NOTE!!! It can also send texts to phones. You need the phone number +
the domain that you send texts to. For examlpe, if you want to send a
message to 555-555-1234 on Verizon, you would send an email to
5555551234@vtext.com