[++] Pure python MD5
This took a few days, because work, life, you know?
The reason I created this is to use a much smaller amount of memory,
because really all I'm using from the Md5 is the hexdigest...
I use the .encode() INSIDE of the Md5 because it gets annoying to have
to type .encode() every single time I want to Md5 something or send
something through a socket (OR BOTH *insert fury here*).