[++] ThreadedTCPServer.py
My server file. The main server SimpleServer starts, and every time a
connection happens, it spawns a UserHandler thread to handle the user.
Userhandler will be the thing that handles everything the user does,
from logging in, activating account, and registering to buying card
packs and viewing your card inventory.
TestHandler is just a handler to test things on without having to modify
UserHandler
sendRecv sends a message (sendmsg) through the socket (socket). if the
type is 'i' then it sends a command which the client interprets as
letting the user input data. recvsize sets the maximum receive data (in
bytes) to that number.
The reason I slice to recvsize is because if you don't, for some reason
the next .recv() call will take the remaining extra bytes from the last
.recv() and it messes. everything. up.
I think I'll be putting sendRecv in the handler later on.