[++] Updated Queues + Added Activation Queue
Queues now work properly. Previously, I had no clue what I was doing
with queues, and my functinos returned True or False to the queue, since
I ran the function in the UserHandler thread and all it did was return
the result to the queue, and the queue worker was just pooping out Trues
and Falses. Now I'm putting the actual things I need to in the queue so
that the QueueWorker can access it and do the function itself. Along
with that, I've added the funct param to QueueWorker.__init__() so that
the queue worker is actually performing the function, rather than
emptying the useless garbage bin queue.