[++] Completed UserHandler.activate() (in UserHandler.register() commit)
I kind of messed up on the commit and put it with the
UserHandler.register(). OOPS!
Activate takes paramaters and activates you a little differently. If you
activate from the main loop in handle(), you will be asked to put in
your username and password and activation code, but if you attempt to
login, the info is correct, and you aren't activated yet, it will get to
the point and ask for your activation code.