[+-] Email thread changed to queue + worker
Removed Email thread functionality, and converted it to an Email Queue
and Email Queue Worker. I also changed the paramaters in
pyemail.send_email to be taken as a tuple\list (so it can work properly
with queue worker) and split up. It's the same thing as before, just 1
more set of parens.