[++] Big update to DoubleLinkedList
All functions referenced will be inside of the DoubleLinkedList class.

* Added __str__ so I can easily print the lists. show will be used for
more of a debugging type processs, as it shows the last, current, and
next reference's data.
* Added __len__ to give a length of how long the Doubly Linked List (but
its a dict you know) is.
* Converted append to _append. _append can append to either side of the
list depending on the start paramater (True is .start side, False is
.end side). This is accessed by appendright and appendleft.
* Added appendright and appendleft. appendright appends data to the
right side of the list, appendleft appends data to the left side of the
list
* Added _extend. This appends each item of the iterable to either side
of the list, depending on the start paramater (True is .start side,
False is .end side). The iterable can be flipped if desired. _extend is
accessed by extendright and extendleft.
* Added extendright and extendleft. The flip paramater on extendleft is
True, because if it is false, it will extend the reversed version of
your iterable (e.g. dllist.extendleft([1,2,3,4,5]) will appear as
[5,4,3,2,1])
* Added remall and start paramaters to remove. The remall paramaters
controls whether or not you want to remove either the first instance or
all of the values that equal that parameter. The start paramater
controls which side of the list you start on (True for .start, False for
.end)