[+-] Failed dequeue, going to rewrite
This dequeue in particular failed. It would not properly remove objects
from the right of the dequeue. It would remove them, but not properly
update all of the other references, therefore would not produce the
right output for __str__. It also would get stuck in loops and not
remove properly.

I will be re-implementing these using a reference class rather than
having .start and .end dictionaries